net-misc/moblock: Fix pkg_postinst
svn path=/sunrise/; revision=5866

+++ net-misc/moblock/moblock-0.8-r1.ebuild
@@ -59,7 +59,7 @@ pkg_postinst() {
 }
 
 pkg_postrm() {
-	if [[ -d /var/cache/moblock ]] ; then
+	if ! has_version ${CATEGORY}/${PN} && [[ -d ${ROOT}/var/cache/moblock ]] ; then
 		einfo "Removing leftover cache..."
 		rm -rf /var/cache/moblock
 	fi
